Add day 6
Tried to use reduce/reduced for all the main processing. It really is such a
versatile function. Together with (range) for input and a vector that you
destructure to have an accumulator consisting of multiple values it's extremely
elegant.

I liked that for part two all I had to do was change the set of history to a
map, and replace conj with assoc. The `contains?` call could stay the same, and
now I can pull out the needed information from the map again in the end.